// src/models/applicationState.ts
export interface IProviderOptions {
    [key: string]: any;
}
// ...
export interface IProject {
    id: string;
    name: string;
    version: string;
    securityToken: string;
    description?: string;
    tags: ITag[];
    sourceConnection: IConnection;
    sourceListConnection: string[];
    targetConnection: IConnection;
    exportFormat: IExportFormat;
    trainFormat: ITrainFormat;
    videoSettings: IProjectVideoSettings;
    activeLearningSettings: IActiveLearningSettings;
    autoSave: boolean;
    assets?: { [index: string]: IAsset };
    lastVisitedAssetId?: string;
}

// region detectron2
export interface ITagDetectron2 {
    id: number;
    name: string;
    isthing: number;
    color: [number, number, number];
}
export interface ICocoCategories {
    things_num: number;
    not_things_num: number;
    categories: ITagDetectron2[];
}
/**
 * 最新的detectron2训练参数
 */
export interface IDetectron2 {
    detectron2Image: string;
    use_gpus: number;
    detectron2: {
        _BASE_: string; // 前端没有设置需要手动
        MODEL: {
            WEIGHTS: string; // 前端没有设置需要手动
            MASK_ON: boolean; // 前端没有设置需要手动
            ROI_HEADS: { NUM_CLASSES: string }; // 前端没有设置需要手动
            RESNETS: { DEPTH: number };
            PIXEL_STD: string;
        };

        SOLVER: {
            MAX_ITER: number;
            STEPS: string; // 前端没有设置需要手动
            IMS_PER_BATCH: string; // 前端没有设置需要手动
        };

        INPUT: {
            MIN_SIZE_TRAIN: string;
            MAX_SIZE_TRAIN: string;
            CROP: {
                ENABLED: string;
                TYPE: string;
                SIZE: string;
            }
        };

        DATALOADER: {
            NUM_WORKERS: string;
            ASPECT_RATIO_GROUPING: string;
            SAMPLER_TRAIN: string;
            REPEAT_THRESHOLD: string;
        };

        DATASETS: {
            TRAIN: string; // 前端没有设置需要手动
            TEST: string; // 前端没有设置需要手动
        };

        OUTPUT_DIR: string; // 前端没有设置需要手动
    };
}
// ...
export interface ITag {
    name: string;
    color: string;
}
// ...
export interface IConnection {
    id: string;
    name: string;
    description?: string;
    providerType: string;
    providerOptions: IProviderOptions | ISecureString;
    providerOptionsOthers?: IProviderOptions[] | ISecureString[]; // 这里新增其他文件夹
}

/**
 * @name - Export Provider Options
 * @description - options defining the type of asset to export
 * @member assetState - export asset with the following state
 */
export interface IExportProviderOptions extends IProviderOptions {
    assetState: ExportAssetState;
}

/**
 * @name - Export Format
 * @description - Defines the settings for how project data is exported into commonly used format
 * @member id - Unique identifier for export format
 * @member name - Name of export format
 * @member providerType - The export format type (TF Records, YOLO, CSV, etc)
 * @member providerOptions - The provider specific option required to export data
 */
export interface IExportFormat {
    providerType: string;
    providerOptions: IExportProviderOptions | ISecureString;
}

/**
 * @name - Video Tagging Settings for the project
 * @description - Defines the video settings within a VoTT project
 * @member frameExtractionRate - Extraction rate for a video (number of frames per second of video)
 */
export interface IProjectVideoSettings {
    frameExtractionRate: number;
}

/**
 * @name - Model Path Type
 * @description - Defines the mechanism to load the TF.js model for Active Learning
 * @member Coco - Specifies the default/generic pre-trained Coco-SSD model
 * @member File - Specifies to load a custom model from filesystem
 * @member Url - Specifies to load a custom model from a web server
 */
export enum ModelPathType {
    Coco = "coco",
    File = "file",
    Url = "url",
}
// ...
export interface IActiveLearningSettings {
    modelPathType: ModelPathType;
    modelPath?: string;
    modelUrl?: string;
    autoDetect: boolean;
    predictTag: boolean;
}

export interface ITrainFormat {
    ip: string;
    providerType: string;
    tarBaseName?: string;
    providerOptions: IYoloV3 | IDetectron | IDetectron2;
}
// ...
export interface IAsset {
    id: string;
    type: AssetType;
    state: AssetState;
    name: string;
    path: string;
    size: ISize;
    format?: string;
    timestamp?: number;
    parent?: IAsset;
    predicted?: boolean;
    tagType?: string;
}

/**
 * @name - Asset Metadata
 * @description - Format to store asset metadata for each asset within a project
 * @member asset - References an asset within the project
 * @member regions - The list of regions drawn on the asset
 */
export interface IAssetMetadata {
    asset: IAsset;
    regions: IRegion[];
    version: string;
}

/**
 * @name - Size
 * @description - Defines the size and/or diminsion for an asset
 * @member width - The actual width of an asset
 * @member height - The actual height of an asset
 */
export interface ISize {
    width: number;
    height: number;
}
// ...
export interface IRegion {
    id: string;
    type: RegionType;
    tags: string[];
    points?: IPoint[];
    boundingBox?: IBoundingBox;
}

/**
 * @name - Bounding Box
 * @description - Defines the tag usage within a bounding box region
 * @member left - Defines the left x boundary for the start of the bounding box
 * @member top - Defines the top y boundary for the start of the boudning box
 * @member width - Defines the width of the bounding box
 * @member height - Defines the height of the bounding box
 */
export interface IBoundingBox {
    left: number;
    top: number;
    width: number;
    height: number;
}

/**
 * @name - Point
 * @description - Defines a point / coordinate within a region
 * @member x - The x value relative to the asset
 * @member y - The y value relative to the asset
 */
export interface IPoint {
    x: number;
    y: number;
}

/**
 * @name - Asset Type
 * @description - Defines the type of asset within a project
 * @member Image - Specifies an asset as an image
 * @member Video - Specifies an asset as a video
 */
export enum AssetType {
    Unknown = 0,
    Image = 1,
    Video = 2,
    VideoFrame = 3,
    TFRecord = 4,
}

/**
 * @name - Asset State
 * @description - Defines the state of the asset with regard to the tagging process
 * @member NotVisited - Specifies as asset that has not yet been visited or tagged
 * @member Visited - Specifies an asset has been visited, but not yet tagged
 * @member Tagged - Specifies an asset has been visited and tagged
 */
export enum AssetState {
    NotVisited = 0,
    Visited = 1,
    Tagged = 2,
    OkTagged = 3,
}

/**
 * @name - Region Type
 * @description - Defines the region type within the asset metadata
 * @member Square - Specifies a region as a square
 * @member Rectangle - Specifies a region as a rectangle
 * @member Polygon - Specifies a region as a multi-point polygon
 */
export enum RegionType {
    Polyline = "POLYLINE",
    Point = "POINT",
    Rectangle = "RECTANGLE",
    Polygon = "POLYGON",
    Pencil = "PENCIL",
    Square = "SQUARE",
}

export enum EditorMode {
    Rectangle = "RECT",
    Polygon = "POLYGON",
    Pencil = "PENCIL",
    Polyline = "POLYLINE",
    Point = "POINT",
    Select = "SELECT",
    CopyRect = "COPYRECT",
    None = "NONE",
}

export interface ISecureString {
    encrypted: string;
}